refactored the application structure to use flask blueprints properly
