+++ app/templates/500_server_error.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% block title %}500 Server Error{% endblock %}
+{% block title %}500 Server Error Flask{% endblock %}
 {% block head %}
 {{ super() }}
 {% endblock %}
