+++ app/website/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("website", __name__)
+
+from app.website import routes
